move "HibernateUtil" to package "util"
